requirements.txt
hydra-core
hydra-colorlog
# torch
torchvision
numpy
scipy
tqdm
tensorboard